requirements.txt
docx==0.2.4
matplotlib==3.1.3
MySQL-python==1.2.5
numpy==1.21.6
opencv_python==4.5.5.64
Pillow==9.2.0
pycocotools==2.0.4
PyQt5==5.15.7
python_docx==0.8.11
scipy==1.4.1
torch==1.9.0+cu111
torchsummary==1.5.1
torchvision==0.10.0+cu111
tqdm==4.42.1
